# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service API",
        "description": "API de gestion des utilisateurs",
        "contact": {
            "email": "n.taffot@elyft.tech"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/apikeys/api-usage-logs/{log}": {
            "get": {
                "tags": [
                    "API Usage Logs"
                ],
                "summary": "Afficher un journal d’usage",
                "operationId": "ca553a37386860cde76197072944da62",
                "parameters": [
                    {
                        "name": "log",
                        "in": "path",
                        "description": "ID du log",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "f4d6eb57-45f8-4b89-9d12-cae4c56877f4"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du journal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}